fix(telemetry): resolve fingerprint for every entry point, not just tool-server
The distinct_id was only fingerprint-derived when the tool-server ran.
The installer/CLI emitted under a random UUID, and uninstall deletes the
id file, so install -> uninstall -> reinstall produced a different random
id each time instead of one stable per-machine id.

Resolve the host fingerprint inside telemetry's buildPayload (the single
shared call site for installer, CLI, tool-server and MCP) so every entry
point derives the same stable id. Move resolveHostFingerprint into
@argent/telemetry (depends on @argent/native-devtools-ios) and drop the
per-caller injection.

Because native-devtools-ios computes its bin dir from __dirname at module
top-level, it must not be inlined into an ESM bundle without a __dirname
shim. Extend the esbuild ESM banner to define __dirname/__filename
(CJS bundles get them natively) so the installer/CLI/MCP ESM bundles
resolve the simulator-server binary instead of crashing on load.
