+++ packages/telemetry/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@argent/configuration-core": "file:../configuration-core",
+    "@argent/native-devtools-ios": "file:../native-devtools-ios",
     "@argent/registry": "file:../registry",
     "ci-info": "^4.4.0",
     "posthog-node": "5.38.6"
